requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==3.12.3
motor==2.3.1
pydantic==2.5.0
python-dotenv==1.0.0
slowapi==0.1.9
bcrypt==4.1.1
python-jose==3.3.0
pydantic-settings==2.1.0
pandas==2.1.3
openpyxl==3.1.5
httpx==0.25.2
